+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - BREAKING (shock!): version signature (adding build meta to pkg version) is turned off by default (use explicit `--sig`)
 - link `.bin` scripts only when `yalc link` or `yalc add --link`
 - read package manifest after running pre-script while publish
+- add `-W` flag to add with `workspace:` protocol
+- fix `--scripts` flag
 
 ## 1.0.0.pre.50 (2021-01-19)